requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
C:/Users/b9739/Downloads/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
altgraph==0.17.2
beautifulsoup4==4.10.0; python_version >= '3.1'
certifi==2021.10.8
charset-normalizer==2.0.9; python_version >= '3'
docopt==0.6.2
et-xmlfile==1.1.0; python_version >= '3.6'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.3; python_version >= '3'
js2py==0.71
openpyxl==3.0.9
packaging==21.3; python_version >= '3.6'
pathlib==1.0.1
pefile==2021.9.3; sys_platform == 'win32'
pipwin==0.5.1
pydantic==1.8.2
pyinstaller-hooks-contrib==2021.4
pyinstaller==4.7
pyjsparser==2.7.1
pynput==1.7.5
pyparsing==3.0.6; python_version >= '3.6'
pyprind==2.11.3
pysmartdl==1.3.4
python-dotenv==0.19.2
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pywin32-ctypes==0.2.0; sys_platform == 'win32'
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.1; python_version >= '3.6'
speechrecognition==3.8.1
sqlalchemy==1.4.28
typing-extensions==4.0.1; python_version >= '3.6'
tzdata==2021.5; platform_system == 'Windows'
tzlocal==4.1; python_version >= '3.6'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
